refactor: extract cookie utilities and add fetch error handling
- Extract readCookie/writeCookie to shared /src/lib/cookies.ts
- Add response.ok checks before parsing JSON in NavItems fetch calls
- Remove duplicate cookie utilities from ThreePanelProvider and NavItems
